+++ controllers/management_cluster_network_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"time"
 
 	"github.com/google/uuid"
 	. "github.com/onsi/ginkgo/v2"
@@ -22,7 +23,9 @@ import (
 
 	"github.com/aws-resolver-rules-operator/controllers"
 	"github.com/aws-resolver-rules-operator/pkg/conditions"
+	gserrors "github.com/aws-resolver-rules-operator/pkg/errors"
 	"github.com/aws-resolver-rules-operator/pkg/k8sclient"
+	"github.com/aws-resolver-rules-operator/pkg/resolver"
 	"github.com/aws-resolver-rules-operator/pkg/resolver/resolverfakes"
 )
 
@@ -54,7 +57,7 @@ var _ = Describe("ManagementClusterTransitGatewayReconciler", func() {
 	BeforeEach(func() {
 		ctx = context.Background()
 
-		cluster = createRandomCluster(
+		_, cluster = createRandomClusterWithIdentity(
 			annotation.NetworkTopologyModeAnnotation,
 			annotation.NetworkTopologyModeGiantSwarmManaged,
 		)
@@ -65,11 +68,14 @@ var _ = Describe("ManagementClusterTransitGatewayReconciler", func() {
 		clusterClient := k8sclient.NewAWSClusterClient(k8sClient)
 		transitGatewayClient = new(resolverfakes.FakeTransitGatewayClient)
 		prefixListClient = new(resolverfakes.FakePrefixListClient)
+		clientsFactory := &resolver.FakeClients{
+			TransitGatewayClient: transitGatewayClient,
+			PrefixListClient:     prefixListClient,
+		}
 		reconciler = controllers.NewManagementClusterTransitGateway(
 			client.ObjectKeyFromObject(cluster),
 			clusterClient,
-			transitGatewayClient,
-			prefixListClient,
+			clientsFactory,
 		)
 	})
 
@@ -273,6 +279,17 @@ var _ = Describe("ManagementClusterTransitGatewayReconciler", func() {
 
 				It("returns an error", func() {
 					Expect(reconcileErr).To(MatchError(ContainSubstring("boom")))
+				})
+
+				When("it isn't detached yet", func() {
+					BeforeEach(func() {
+						transitGatewayClient.DeleteReturns(gserrors.NewRetryableError("boom", time.Second))
+					})
+
+					It("requeues the event", func() {
+						Expect(reconcileErr).NotTo(HaveOccurred())
+						Expect(reconcileResult.RequeueAfter).To(Equal(time.Second))
+					})
 				})
 			})