+++ pkg/aws/errors.go
@@ -2,10 +2,41 @@ package aws
 
 import (
 	"reflect"
+
+	"github.com/aws/aws-sdk-go/aws/awserr"
+	"github.com/pkg/errors"
 )
 
-type SecurityGroupNotFoundError struct {
+const (
+	ErrAssociationNotFound = "InvalidAssociationID.NotFound"
+	ErrRouteTableNotFound  = "InvalidRouteTableID.NotFound"
+	ErrPrefixListNotFound  = "InvalidPrefixListID.NotFound"
+	ErrSubnetNotFound      = "InvalidSubnetID.NotFound"
+	ErrVPCNotFound         = "InvalidVpcID.NotFound"
+	ErrIncorrectState      = "IncorrectState"
+)
+
+func HasErrorCode(err error, code string) bool {
+	var apiError awserr.Error
+	ok := errors.As(err, &apiError)
+	if !ok {
+		return false
+	}
+
+	return apiError.Code() == code
 }
+
+type TransitGatewayNotReadyError struct{}
+
+func (e *TransitGatewayNotReadyError) Error() string {
+	return "transit gateway not ready"
+}
+
+func (e *TransitGatewayNotReadyError) Is(target error) bool {
+	return reflect.TypeOf(target) == reflect.TypeOf(e)
+}
+
+type SecurityGroupNotFoundError struct{}
 
 func (e *SecurityGroupNotFoundError) Error() string {
 	return "security group was not found"
@@ -15,8 +46,7 @@ func (e *SecurityGroupNotFoundError) Is(target error) bool {
 	return reflect.TypeOf(target) == reflect.TypeOf(e)
 }
 
-type ResolverEndpointNotFoundError struct {
-}
+type ResolverEndpointNotFoundError struct{}
 
 func (e *ResolverEndpointNotFoundError) Error() string {
 	return "resolver endpoint was not found"
@@ -26,8 +56,7 @@ func (e *ResolverEndpointNotFoundError) Is(target error) bool {
 	return reflect.TypeOf(target) == reflect.TypeOf(e)
 }
 
-type HostedZoneNotFoundError struct {
-}
+type HostedZoneNotFoundError struct{}
 
 func (e *HostedZoneNotFoundError) Error() string {
 	return "hosted zone was not found"
@@ -37,8 +66,7 @@ func (e *HostedZoneNotFoundError) Is(target error) bool {
 	return reflect.TypeOf(target) == reflect.TypeOf(e)
 }
 
-type DnsRecordNotSupportedError struct {
-}
+type DnsRecordNotSupportedError struct{}
 
 func (e *DnsRecordNotSupportedError) Error() string {
 	return "dns record type is not supported"
@@ -48,8 +76,7 @@ func (e *DnsRecordNotSupportedError) Is(target error) bool {
 	return reflect.TypeOf(target) == reflect.TypeOf(e)
 }
 
-type RouteTableNotFoundError struct {
-}
+type RouteTableNotFoundError struct{}
 
 func (e *RouteTableNotFoundError) Error() string {
 	return "route table was not found"