+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `TransitGatewayAttachmentReconciler`.
 - Add `PrefixListEntryReconciler`.
 - Add `ShareReconciler`.
+- Add acceptance tests.
+
 ### Changed
 
 - Configure `gsoci.azurecr.io` as the default container image registry.